feat: improve navigation responsiveness in the interviews page
