+++ app/dashboard/_components/InterviewsTable/InterviewsTableClient.tsx
@@ -1,0 +1,115 @@
+'use client';
+
+import { use, useMemo, useState } from 'react';
+import { ActionsDropdown } from '~/app/dashboard/_components/InterviewsTable/ActionsDropdown';
+import { InterviewColumns } from '~/app/dashboard/_components/InterviewsTable/Columns';
+import { DataTable } from '~/components/DataTable/DataTable';
+import { Button } from '~/components/ui/Button';
+import { DeleteInterviewsDialog } from '~/app/dashboard/interviews/_components/DeleteInterviewsDialog';
+import { ExportInterviewsDialog } from '~/app/dashboard/interviews/_components/ExportInterviewsDialog';
+import { HardDriveUpload } from 'lucide-react';
+import { GenerateInterviewURLs } from '~/app/dashboard/interviews/_components/GenerateInterviewURLs';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '~/components/ui/dropdown-menu';
+import type { GetInterviewsReturnType } from '~/queries/interviews';
+import type { GetProtocolsReturnType } from '~/queries/protocols';
+
+const InterviewsTableClient = ({
+  interviewsPromise,
+  protocolsPromise,
+}: {
+  interviewsPromise: GetInterviewsReturnType;
+  protocolsPromise: GetProtocolsReturnType;
+}) => {
+  const interviews = use(interviewsPromise);
+
+  const [selectedInterviews, setSelectedInterviews] =
+    useState<typeof interviews>();
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
+  const [showExportModal, setShowExportModal] = useState(false);
+
+  const unexportedInterviews = useMemo(
+    () => interviews.filter((interview) => !interview.exportTime),
+    [interviews],
+  );
+
+  const handleDelete = (data: typeof interviews) => {
+    setSelectedInterviews(data);
+    setShowDeleteModal(true);
+  };
+
+  const handleExportUnexported = () => {
+    setSelectedInterviews(unexportedInterviews);
+    setShowExportModal(true);
+  };
+
+  const handleExportAll = () => {
+    setSelectedInterviews(interviews);
+    setShowExportModal(true);
+  };
+
+  const handleResetExport = () => {
+    setSelectedInterviews([]);
+    setShowExportModal(false);
+  };
+
+  return (
+    <>
+      <ExportInterviewsDialog
+        key={selectedInterviews?.toString()}
+        open={showExportModal}
+        handleCancel={handleResetExport}
+        interviewsToExport={selectedInterviews!}
+      />
+      <DeleteInterviewsDialog
+        open={showDeleteModal}
+        setOpen={setShowDeleteModal}
+        interviewsToDelete={selectedInterviews ?? []}
+      />
+      <DataTable
+        columns={InterviewColumns()}
+        data={interviews}
+        filterColumnAccessorKey="identifier"
+        handleDeleteSelected={handleDelete}
+        handleExportSelected={(selected) => {
+          setSelectedInterviews(selected);
+          setShowExportModal(true);
+        }}
+        actions={ActionsDropdown}
+        headerItems={
+          <>
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button disabled={interviews.length === 0}>
+                  <HardDriveUpload className="mr-2 inline-block h-4 w-4" />
+                  Export Interview Data
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent>
+                <DropdownMenuItem onClick={handleExportAll}>
+                  Export all interviews
+                </DropdownMenuItem>
+                <DropdownMenuItem
+                  disabled={unexportedInterviews.length === 0}
+                  onClick={handleExportUnexported}
+                >
+                  Export all unexported interviews
+                </DropdownMenuItem>
+              </DropdownMenuContent>
+            </DropdownMenu>
+            <GenerateInterviewURLs
+              interviews={interviews}
+              protocolsPromise={protocolsPromise}
+            />
+          </>
+        }
+      />
+    </>
+  );
+};
+
+export default InterviewsTableClient;
