+++ app/dashboard/interviews/page.tsx
@@ -1,24 +1,13 @@
 import ResponsiveContainer from '~/components/ResponsiveContainer';
 import Section from '~/components/layout/Section';
-import PageHeader from '~/components/ui/typography/PageHeader';
-import { requireAppNotExpired } from '~/queries/appSettings';
-import { requirePageAuth } from '~/utils/auth';
-import InterviewsTableServer from '../_components/InterviewsTable/InterviewsTableServer';
+import InterviewsTable from '../_components/InterviewsTable/InterviewsTable';
 
-export default async function InterviewPage() {
-  await requireAppNotExpired();
-  await requirePageAuth();
+export default function InterviewPage() {
   return (
     <>
-      <ResponsiveContainer>
-        <PageHeader
-          headerText="Interviews"
-          subHeaderText="View and manage your interview data."
-        />
-      </ResponsiveContainer>
       <ResponsiveContainer maxWidth="7xl">
         <Section>
-          <InterviewsTableServer />
+          <InterviewsTable />
         </Section>
       </ResponsiveContainer>
     </>
